Extract default context menu module
